Try to resolve cmsisnn.lib dependency issue (not working)
Currently mlonmcu is not able to figure out the correct patch to the `cmsisnn.lib`. This is due to REQUIRED list of the feature which is resolved in the very beginning before a target was initialized. While the cache flags `x86`/`arm`/`riscv` would be required for looking CMSIS-NN, there is currently no way to provide this additional flag at this point in time.